pyproject.toml
[tool.poetry]
name = "agriculture-backend"
version = "0.1.0"
description = ""
authors = ["Vladislav Stukov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
celery = {extras = ["redis"], version = "^5.4.0"}
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
